Fix: SeatPage 내 유저 아이디 상태 공유, 호출 버튼 구현

+++ Frontend/moass/lib/model/device_info.dart
@@ -3,8 +3,8 @@ import 'package:moass/model/user_info.dart';
 class DeviceInfo {
   final String deviceId;
   final String? userId;
-  final double? xcoord;
-  final double? ycoord;
+  final int? xcoord;
+  final int? ycoord;
   final String classCode;
   final UserInfo? userSearchDetail;
